+++ crates/vertigo-macro/src/html_parser.rs
@@ -114,6 +114,10 @@ fn convert_node(node: Node, convert_to_dom_node: bool) -> Result<TokenStream2, (
             out_attr.push(quote!{
                 .on_change(#value)
             })
+        } else if name == "on_blur" {
+            out_attr.push(quote!{
+                .on_blur(#value)
+            })
         } else if name == "on_key_down" {
             out_attr.push(quote!{
                 .on_key_down(#value)